splitted classes into two seperated files for each class
