+++ ProductManager.mjs
@@ -1,12 +1,4 @@
-export class Product {
-    constructor(name, description, category, amount) {
-        this.name = name || "N/A";
-        this.description = description || "N/A";
-        this.category = category || "N/A";
-        this.amount = amount || 0;
-    }
-}
-
+import { Product } from './Product.mjs';
 import pkg from 'mongodb';
 const { MongoClient, MongoError } = pkg;
