+++ app.mjs
@@ -1,4 +1,4 @@
-import { ProductsManager } from './classes.mjs';
+import { ProductsManager } from './ProductManager.mjs';
 import express from 'express';
 import bodyParser from 'body-parser';
